fix: bug hunt passe 1 — 24 correctifs, dont 2 crashs/évasions critiques
Chasse multi-agents (8 zones disjointes, méthode CLAUDE.md : Hatari lu comme
source de vérité AVANT toute conclusion). Étalons pixel verts avant/après.

CRITIQUES (prouvés par repro, puis re-testés corrigés)
- GemdosHd : ÉVASION du dossier monté. « / » n'était pas reconnu comme
  séparateur GEMDOS (seul « \ » l'était), donc « ../X » traversait le parseur
  intact et ressortait tel quel dans le chemin hôte, où le « .. » était résolu
  pour de bon : un .TOS hostile lisait, ÉCRIVAIT et listait hors du bac à sable,
  avec les droits de l'utilisateur. Deux couches : normalisation du séparateur
  à l'entrée + garde-fou terminal (canonicalisation, préfixe séparateur inclus,
  rabattement sur la racine et journalisation). Vérifié : le fichier créé hors
  montage atterrit désormais dedans, la lecture ne rend plus que des zéros.
  Hatari a le même trou (add_path_component) — durcissement assumé, NeoST étant
  une « boîte à hack » où l'on lance des binaires inconnus.
- Blitter : RÉCURSION INFINIE → SIGSEGV. Les accès bus du blitter passent par
  Bus::write16, qui re-route $FF8A00-3F vers Blitter::write16 : un blit dont la
  destination est son propre registre de contrôle se relançait en plein
  transfert, compteurs non réécrits. Garde de ré-entrance (Hatari ne redémarre
  jamais depuis un accès bus : blitter.c:1421 diffère). Repro 68000 : core dump
  avant, « survecu » après.

MAJEURS
- main.cpp : lecture hors bornes de g_kioskDisks (3 appels sur 4 à
  kioskScanDisks remettaient l'index à 0, le 4e l'oubliait) ; la disquette
  insérée par un visiteur salissait cfg et fuyait au prochain saveConfig(force),
  violant l'invariant « la borne repart identique ».
- Bus::dmaWrite8 laissait le DMA écraser $0-$7 (miroir SSP/PC) — Hatari faute
  inconditionnellement (memory.c:775).
- Shifter : prefetch et line-offset STE appliqués même sur une ligne NON
  affichée, alors que Hatari les garde dans la branche « ligne non blanche »
  (video.c:3993→4213) → dérive du compteur vidéo sur STE à scrolling matériel.
- Fdc : détection .DIM par contenu trop laxiste ($4242 = « clr.w d2 », début
  plausible d'un secteur de boot) → une .ST légitime montée amputée de 32 o,
  décalée et en lecture seule. Champs de géométrie désormais validés.
- GemdosHd : dossier monté gardé RELATIF → tout Dsetpath échouait en EPTHNF.
- headless : --sound-dump + gros --frames = std::bad_alloc (reserve non borné) ;
  --load-state en échec rendait rc=0, faisant passer un oracle au vert sur un
  boot à froid.

EMPAQUETAGE (aucun n'était détectable par le run CI vert)
- polices absentes des 4 paquets → GUI replié en silence sur la police bitmap,
  tous les pictogrammes en carrés vides ;
- version : define figée sur project() → tous les paquets s'annonçaient 0.1.0 ;
- macOS sans CMAKE_OSX_DEPLOYMENT_TARGET → .dmg refusé sous macOS 15, donc sur
  tout Mac Intel d'avant 2019 (le public même de la tranche x86_64) ;
- EmuTOS 256 Ko absent du zip WASM → mode STE silencieusement inopérant ;
- glob effaçant le paquet Raspberry ; CFBundleVersion non conforme.
Gardes CI ajoutées pour chacun (polices, disques non libres, version binaire,
GLIBCXX statique, plancher macOS).

MINEURS : liveTop/snapLiveTop codaient le début d'affichage 50 Hz en dur ;
latches du zoom adaptatif gelés quand il est coupé ; F10 ne resynchronisait pas
la config ; ACIA MIDI jamais réinitialisée (reset.c:111/124) ; $FF8001 non remis
à 0 au froid ; DriveSound avançait son accumulateur de want au lieu de got ;
CrtEffectStack bindait sur l'unité de texture de l'appelant ; StePads décalait
un négatif (UB).

+++ src/io/StePads.hpp
@@ -120,7 +120,7 @@ public:
         if ((select_ & 0xF0) != 0xF0) {
             nData &= 0x0F;
             if (!(select_ & 0x10))                   // ligne « directions »
-                nData |= uint16_t((~stickData(padB_) << 4) & 0xF0);
+                nData |= uint16_t((~unsigned(stickData(padB_)) << 4) & 0xF0);   // unsigned : décaler un négatif est UB
             else nData |= 0xF0;
         }
